% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boosting.R
\name{RealAdaBoostClass}
\alias{RealAdaBoostClass}
\title{Real Adaboost}
\usage{
RealAdaBoostClass(
  mDataTrain = NA,
  mDataTest = NA,
  iM = 10,
  iDepth = 2,
  iSeed = 1,
  bVerbose = TRUE
)
}
\arguments{
\item{mDataTrain}{matrix containig training data, on which weak learners are fitted.}

\item{mDataTest}{matrix containing test data, used for prediction and evaluation.}

\item{iM}{integer, total number of included weak learners, default is 10.}

\item{iDepth}{integer, maximal depth weak learners, default is 2.}

\item{iSeed}{integer, random seed used to do weighted sampling in each iteration step.}

\item{bVerbose}{boolean, prints weak learners when TRUE}
}
\value{
A list containing class predictions, misclassification rate and a list including the base learners.
}
\description{
Real Adaboost Classification based on Friedman, Hastie and Tibshirani (2000)
}
\author{
Ramon de Punder (University of Amsterdam, Tinbergen Institute)
}
