% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boosting.R
\name{RealAdaBoostClass_CV}
\alias{RealAdaBoostClass_CV}
\title{Real Adaboost Cross Validation}
\usage{
RealAdaBoostClass_CV(
  lCVData = NA,
  lArgs = NA,
  vM = seq(1, 15),
  sName = "",
  bPlot = TRUE
)
}
\arguments{
\item{lCVData}{list, has length equal to number of folds, each list element lists a train and test matrix of data}

\item{lArgs}{list, contains other parameters needed for RealAdaBoostClass than to be tuned by CV: Depth and Seed.}

\item{vM}{vector, grid for M, the total number of included weak learners}

\item{sName, }{string, characteristic part file name}

\item{bPlot, }{boolean, creates plot and saves tikzfile in folder Figures when true.}
}
\value{
CV value for M, optionally together with corresponding standard error when bPlot is FALSE.
}
\description{
Cross Validation for Real Adaboost Classification based on Friedman, Hastie and Tibshirani (2000)
}
